General improvements, most notably in the Deskewing functionality. Instead of taking the average of the angles selected by the deskewer, the highest voted angle is selected.

+++ src/imageprocessing/deskew/HoughCoordinates.java
@@ -61,7 +61,7 @@ public class HoughCoordinates
       return mAngle;
    }
 
-   public double getAR()
+   public int getAR()
    {
       return mAngleRef;
    }
